Changed type from int to size_t

+++ pointers_arrays_strings/7-puts_half.c
@@ -7,7 +7,7 @@
 void puts_half(char *str)
 {
 	size_t i; 
-	int lengthOfString;
+	size_t lengthOfString;
 
 	lengthOfString = strlen(str);
